feat(nvim,zshrc): add daily notes keymap and alias
This adds a new keymap in Neovim for quickly opening today's daily note
by pressing `<leader>ot`.

It also adds a new alias to the .zshrc file, `today`, that also opens
today's daily note. To affect this change, lazy loading needed to be
turned off for the Obsidian Neovim plugin.